1.完成第一版爬蟲，可以爬回資料並正確進行到價判斷

+++ Crawler/Craw/BankOfTaiwan.cs
@@ -71,8 +71,8 @@ namespace Crawler.Craw
                 DateTime.TryParse(pCols[0].InnerText, out tmpTime);
                 p.Time = tmpTime;
 
-                double tmpPrice;
-                double.TryParse(pCols[1].InnerText, out tmpPrice);
+                decimal tmpPrice;
+                decimal.TryParse(pCols[1].InnerText, out tmpPrice);
                 p.Price = tmpPrice;
 
                 double tmpRiseValue;
